# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  docs.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
      success:
        type: boolean
    type: object
...
  docs.FoodResponse:
    properties:
      calories:
        type: number
      carbohydrates:
        type: number
      category:
        type: string
      fat:
        type: number
      fiber:
        type: number
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      protein:
        type: number
      sodium:
        type: number
      sugar:
        type: number
    type: object
  docs.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
  docs.MessageResponse:
    properties:
      message:
        type: string
    type: object
  docs.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  docs.RecommendationResponse:
    properties:
      applied_rules:
        items:
          type: string
        type: array
      pagination:
        properties:
          limit:
            type: integer
          offset:
            type: integer
        type: object
      recommendations:
        items:
          $ref: '#/definitions/docs.FoodResponse'
        type: array
      total_count:
        type: integer
    type: object
...
  docs.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        type: string
    type: object
  docs.Response:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
  docs.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@nutrimatch.com
    name: API Support
    url: http://www.nutrimatch.com/support
  description: A personalized nutrition recommendation system built on the OpenNutrition
    dataset
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NutriMatch API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return access and refresh tokens
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/docs.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  $ref: '#/definitions/docs.TokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      summary: Login user
      tags:
      - auth
...
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User registration information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/docs.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  $ref: '#/definitions/docs.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      summary: Register new user
      tags:
      - auth
  /api/v1/foods:
    get:
      consumes:
      - application/json
      description: Search for foods by name or other criteria
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/docs.FoodResponse'
                  type: array
                meta:
                  $ref: '#/definitions/docs.PaginationMeta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      summary: Search foods
      tags:
      - foods
...
  /api/v1/recommendations:
    get:
      consumes:
      - application/json
      description: Get personalized food recommendations for the authenticated user
      parameters:
      - default: 10
        description: Number of recommendations to return
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Profile ID to use for recommendations
        in: query
        name: profileId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  $ref: '#/definitions/docs.RecommendationResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get food recommendations
      tags:
      - recommendations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"